return double from xml_get_prop_float()
Some callers will store the value actually in a double, so do not lose
precision. Those callers that actually use float can narrow it themselves.

+++ src/misc.c
@@ -42,9 +42,9 @@
 #include <string.h>
 #include <strings.h>
 
-float xml_get_prop_float(xmlNode *node, const char *prop) {
+double xml_get_prop_float(xmlNode *node, const char *prop) {
   xmlChar *str = xmlGetProp(node, BAD_CAST prop);
-  float value = NAN;
+  double value = NAN;
   if(str) {
     value = g_ascii_strtod((gchar*)str, NULL);
     xmlFree(str);
